Adds ability to move the camera via the mouse scroll wheel.
Also renames Camera -> WesnothCamera to avoid confusion with Godot features like GetCamera() or Godot.Camera.
